pyproject.toml
[tool.poetry]
name = "misskey-tweet-proxy"
version = "1.0.0"
description = "MisskeyAPIを用いてMisskeyにTweetを送信します"
authors = ["code-raisan <yuri@rainyan.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"Misskey.py" = "^4.1.0"
tweety-ns = "^0.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
